Refactor : 패키지 생성/파일이동 - controller, service 패키지 내 seller, coupon, buyer 패키지 분기
