+++ src/main/java/com/lotte/danuri/product/repository/CategoryFirstRepository.java
@@ -1,6 +1,6 @@
 package com.lotte.danuri.product.repository;
 
-import com.lotte.danuri.product.model.CategoryFirst;
+import com.lotte.danuri.product.model.entity.CategoryFirst;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryFirstRepository extends JpaRepository<CategoryFirst, Long> {
