+++ src/main/java/com/lotte/danuri/product/repository/CategorySecondRepository.java
@@ -1,6 +1,6 @@
 package com.lotte.danuri.product.repository;
 
-import com.lotte.danuri.product.model.CategorySecond;
+import com.lotte.danuri.product.model.entity.CategorySecond;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategorySecondRepository extends JpaRepository<CategorySecond, Long> {
