+++ src/main/java/com/lotte/danuri/product/repository/CategoryThirdRepository.java
@@ -1,6 +1,6 @@
 package com.lotte.danuri.product.repository;
 
-import com.lotte.danuri.product.model.CategoryThird;
+import com.lotte.danuri.product.model.entity.CategoryThird;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryThirdRepository extends JpaRepository<CategoryThird, Long> {
