+++ src/main/java/com/lotte/danuri/product/repository/StoreRepository.java
@@ -1,6 +1,6 @@
 package com.lotte.danuri.product.repository;
 
-import com.lotte.danuri.product.model.Store;
+import com.lotte.danuri.product.model.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
