+++ src/main/java/com/lotte/danuri/product/service/seller/SellerProductService.java
@@ -1,9 +1,9 @@
-package com.lotte.danuri.product.service;
+package com.lotte.danuri.product.service.seller;
 
-import com.lotte.danuri.product.model.Product;
+import com.lotte.danuri.product.model.entity.Product;
 import com.lotte.danuri.product.model.dto.ProductDto;
 
-public interface ProductService {
+public interface SellerProductService {
     ProductDto createProduct(ProductDto productDto);
     Iterable<Product> getAllProducts();
